Prod

+++ client/app/components/tools/Chart.tsx
@@ -62,7 +62,7 @@ export default function MyChart({onDataUpdate}: any) {
   datasets: [
     {
       label: 'Growth Yeild',
-      data: chartData.values.map(value => parseInt(value)),
+      data: chartData.values.map(value => parseFloat(value)),
       backgroundColor:  ['#FF6384', '#36A2EB', '#FFCE56', '#4BC0C0', '#9966FF', '#FF9F40', '#00BFFF', '#FFD700', '#32CD32', '#8A2BE2'],
       barPercentage: 0.5,
     },
